// src/api/nexuscore/model/guildMonthlyPlaytime.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface GuildMonthlyPlaytime {
  /** The month this data is about. Always the first day of that month */
  month: string;
  /** The total playtime that month in seconds */
  total: number | null;
  /** How many unique players played that month */
  unique_players: number;
}