// src/api/nexuscore/model/guildDailyPlaytime.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface GuildDailyPlaytime {
  /** The day this data is about */
  day: string;
  /** The total playtime that day in seconds */
  total: number | null;
  /** How many unique players played that day */
  unique_players: number;
  /** The total amount of sessions that day. (A session is when a user connects and disconnects) */
  total_sessions: number;
  /** The average playtime per session today in seconds */
  average_playtime_per_session: number | null;
}